Archimedes screw version
The original idea was to use a servo motor to open a valve and then a vibration motor to shake the coffee out as it tends to stick together by friction making it unable to fall out of the valve opening. The problem with this setup was that it looked ugly and the valve servo was affected by the vibration motor. When the motor was running the servo lost control of its position and jumped back and forth. I think it was because the wires for the servo and engine were "long" and thread next to each other. This might have induced signals in the control wire for the servo. Just a theory. 

This version replaces the coffee container with a prettier solution and replaces the servo/vib motor combo with only one continuously rotating servo. The noise problem is removed and the look is better.

The opening angle of the filter has been tweaked to accomodate this change. Now two open positions exist. One for filling the filter with coffee and one for changing the filter. This is changed in the class definition of the coffee machine and all external functions work in the same way except for fillCoffee which now require an int, 1 or 2, for deciding which position to open to.